python/requirements.txt
asyncio==3.4.3
bidict==0.23.1
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
colorzero==2.0
distro==1.8.0
flake8==7.3.0
gpiozero==2.0.1
h11==0.16.0
idna==3.3
keyboard==0.13.5
lgpio==0.2.2.0
mccabe==0.7.0
pigpio==1.78
pycodestyle==2.14.0
pycryptodomex==3.11.0
pyflakes==3.4.0
python-engineio==4.12.2
python-socketio[client]==5.13.0
requests==2.31.0
rpi-lgpio==0.6
simple-websocket==1.1.0
six==1.16.0
smbus2==0.4.2
spidev==3.5
ssh-import-id==5.10
toml==0.10.2
urllib3==1.26.12
wsproto==1.2.0
websocket-client
aiohttp
opencv-python
json